+++ src/kth/csc/inda/pockettheremin/synth/SoundGenerator.java
@@ -1,12 +1,12 @@
 package kth.csc.inda.pockettheremin.synth;
 
 public interface SoundGenerator {
-	public static double FREQUENCY_MAX = 20000; // Hz
-	public static double FREQUENCY_MIN = 1; // Hz
+	public static double FREQUENCY_MAX = 20000.00; // Hz
+	public static double FREQUENCY_MIN = 1.00; // Hz
 
 	public short[] getSamples(int amount);
 
-	public void setSampleRate(int sampleRate);
+	public void setSampleRate(double sampleRate);
 
 	public void setFrequency(double frequency);
 }
